Add bookings

+++ src/main/java/ru/practicum/shareit/other/exception/AccessDeniedForUserException.java
@@ -1,7 +1,7 @@
 package ru.practicum.shareit.other.exception;
 
 public class AccessDeniedForUserException extends RuntimeException {
-    public AccessDeniedForUserException(int userId) {
+    public AccessDeniedForUserException(long userId) {
         super("Access for user '" + userId + "' is denied");
     }
 }
